pyproject.toml
[project]
name = "mail_parser"
version = "1.0.0"
description = "High-performance Gmail mbox parser with multi-dimensional organization and analytics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "tqdm>=4.66.0",
    "chardet>=5.2.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "click>=8.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mail_parser"]

[tool.hatch.build]
include = ["mail_parser/**/*.py", "mail_parser/**/*.html"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]
